Added web workers for auto generating art. Renamed deepLearning.js to otherEffects.js
Features:
When generating art, a loading screen is displayed.
Once done, the data is reused over and over, so the loading screen is not needed, unless a new image is uploaded.

Also moved almost everthing but setup and draw from poly.js to polyElse.js. This makes it easier for webWorkerArtGen.js to load scripts without having to load p5.js, which for some reason doesn't work.

Removed some deprecated variables

Renamed to otherEffects.js because it is literally not about deep learning