Update Poly.js
Massive Performance Upgrade
Triangles are now drawn on a off screen canvas, saved as an image, and then drawn on the on screen canvas, making it easier to edit the poly art as you go.

After downloading image, canvas doesn't get super large and unusable.

Flowering effect is kept on a seprate js file.

Triangle flipper is currently broken and won't be used.